Add documentation to some modules and change the folder structure
