+++ frontend/src/js/util/histogram.js
@@ -1,4 +1,12 @@
-// Returns a histogram for the entries of the input list.
+/**
+ * Compute a histogram for the entries of the input list.
+ *
+ * @param {[String]} l A list of strings containing the values for which a
+ *                     histogram is computed.
+ *
+ * @returns            An object with the list elements as properties and the
+ *                     list element/number of occurrences pair as properties.
+ */
 export function histogram(l) {
   const histogram = {};
   l.forEach((e) => {
@@ -10,8 +18,13 @@ export function histogram(l) {
   return histogram;
 }
 
-// Returns the value that occurs in the input list most often, or null
-// for an empty list.
+/**
+ * Compute the element that occurs in the input list most often.
+ *
+ * @param {[String]} l
+ * @returns the element occurring most often, or @null for an empty input list.
+ *          If more than one element occurs the most often, the ifrst is chosen
+ */
 export function majorityValue(l) {
   if (l.length === 0) return null;
   const histo = Object.values(histogram(l));
